WIP
So at this point, I realize that _Pipe being iso is needed for this
approach but also problematic because they also need to go to _Process.

I think I need to find a way to merge the ProcessMonitorCreator and
_Process into something that makes sense.

Or I could give ownership of the pipes entirely to _Process and
interact with them as needed from the monitor through the _Process.