Revert "Fix issue that could lead to a muted actor being run (#3974)"
This reverts commit bc927a619a139cd1270a723613652b4506005ae1.
I realized when looking at #4069 tat there is no edge-case. I had
forgotten when this commit was authored that an actor's queue isn't
"empty" until we mark it empty. Therefore, no edge case.

The issue this was attempting to fix is the same issue as #4069 which
I am currently testing a fix for.

Closes #4076