Revert "Fix issue that could lead to a muted actor being run (#3974)"
This reverts commit bc927a619a139cd1270a723613652b4506005ae1.
When I identified the edge case, I forgot that when an actor gets
muted, it returns without marking its queue as empty and therefore,
won't be seen as being empty if another actor sends a message to it
even if there aren't actually messages in queue.

The error that causes the assertion failure that led me to doing the
original commit is probably the same that is causing #4069 and when
that is fixed, the assertion for a muted actor being run will also
be fixed.

Closes #4076