Strengthen the ordering for some atomic operations (#4083)
This commit undoes the changes introduced in 2016 in 
https://github.com/ponylang/ponyc/pull/1209. 

While we don't know that every change in #1209, we do know that a number of the
changes are the source of incorrect operations on Arm where we see evidence of
"bad things" like actors being run by more than 1 scheduler at a time.

Matt Parkinson from MSR Cambridge identified a number of issues with compare
and swap on Arm that Joe Eli McIlvain then found were introduced in #1209. The
issues would not impact x86, only CPUs with weaker memory models like Arm. 

Given that we know that some of the changes in #1209 are wrong, Joe and I 
decided that the best course of action is to roll them all back. While all 
might not need to be rollback, we are following the Pony philosophy and valuing
correctness over performance.

We have much better testing now then we did 5 years ago and can put any future
relaxing of atomic order operations through a considerably more vigorous testing
process than we did years ago for #1209.

With these sorts of bugs, we can never say, they have been fixed. What we can 
say is that we ran with Matt's suggested changes for over 48 hours without
encountering any crashes or other signs of incorrect behavior. As such, we 
believe that these changes are an improvement but our nightly runtime stress
tests might find additional problems in the future.

It should be noted, that due to changes in the last 5 years, this is not an 
exact reversion. I've done my best to make sure that all changes were reverted
as they would be seen in their current form today.

Closes #4069