Fix initialization of an atomic variable in `ponyassert.c`.
While working on cross-compiling the runtime from Linux to Windows,
I ran into this issue which was causing clang to fail with:

```
src/libponyrt/platform/ponyassert.c:19:26: error: copying variable of type 'std::atomic<bool>' invokes deleted constructor
static PONY_ATOMIC(bool) assert_guard = false;
                         ^              ~~~~~
/tmp/xwin/crt/include/atomic:2276:5: note: 'atomic' has been explicitly marked deleted here
    atomic(const atomic&) = delete;
    ^
```

Apparently the correct way to initialize a `std::atomic` is with
an initializer list, as described in this blog post:
https://onetwobytes.wordpress.com/2021/02/24/std-atomic-defaults/

However, we don't actually use `std::atomic` on all platforms,
and the platforms that don't use it need to keep using the `=` syntax.
So the initializer needs to get moved into a new dedicated macro,
which I have named `PONY_ATOMIC_INIT` and used for this purpose.