Reinstate `noalias` attribute for all but the `alloc_final` functions.
Only the `pony_alloc_final*` functions need to avoid the `noalias`
attribute, because running a finaliser is the only case where we
have the issue where the runtime needs to dereference memory with
the alias held by the object map, but LLVM sees that the object
was never read from again, and thus can optimize it away.
Only finaliser effects prevent this LLVM optimization.