Fix compilation error with PONY_NDEBUG
`frozen` is only available on ast and token if PONY_NDEBUG isn't defined.
However, it was used in two functions regardless of the value of PONY_NDEBUG.

This isn't caught in CI as we currently use PONY_ALWAYS_ASSERT which means that
PONY_NDEBUG isn't on for any of our CI (or any other builds).