Don't always assert
This removes the inclusion of PONY_ALWAYS_ASSERT when building and
adds the small changes needed to compile with that change.

We haven't been running with PONY_NDEBUG on for a long time as
PONY_ALWAYS_ASSERT "turns off" PONY_NDEBUG. This means that for the
most part, config=release has result in debug versions of the runtime.