Don't use "mostly debug runtime" with release configurations (#4112)
This removes the inclusion of PONY_ALWAYS_ASSERT when building and
adds the small changes needed to compile with that change.

We haven't been running with PONY_NDEBUG on for a long time as
PONY_ALWAYS_ASSERT "turns off" PONY_NDEBUG. This means that for the
most part, config=release has result in debug versions of the runtime.

This commit also rearranges our PR CI order so that `config=debug` comes
before `config=release`. If we do introduce a compiler error, it is better
to trigger an assertion than a crash. We stop running CI after the first error
so we could be left with a segfault that would be explained by an assertion.
The reordering so that debug is run before release assures that we will see
the most informative error messages.

The commit also fixes an assert in refer.c that had a side-effect. Compiling ponyc with
assertions off would lead to a compiler that segfaulted because a required ast_pop
wasn't being done.

The commit includes many small changes to handle variables that are only used for
assertions and lead to compilation errors.