Disable incorrect runtime assert for ASIO thread shutdown.
When shutting down the ASIO thread and destroying its message queue,
it is possible that there may be some message(s) in flight to it,
meaning that even though we drained the queue before destroying,
the queue may be non-empty by the time we get to destroying it.
Hence, we cannot reliably assert that the queue is non-empty.

However, for actors, we can proved definitively that an actor
is done receiving messages before we destroy its message queue,
so we want to retain the runtime assert for the actor case.

This change updates the assert to be conditional based on
a boolean parameter to the `ponyint_messageq_destroy` function.

For more context, see [this Zulip thread](https://ponylang.zulipchat.com/#narrow/stream/190365-runtime/topic/ponyint_messageq_destroy.20assert.20fail).