Remove `actor->is_muted` and replace with a flag
`actor->is_muted` is a `uint8_t` but it is always treated as a `boolean`.
It is also always only modified from a single scheduler thread (either to mute
or unmute the actor) but read from other scheduler threads. This is exactly
the same behavior as that of the various `actor->flags`.

This commit removes `actor->is_muted`, extends `actor->flags` from
`uint8_t` to `uint16_t` and adds a new flag called `FLAG_MUTED`. The
logic in `ponyint_is_muted`, `ponyint_mute_actor` and `ponyint_unmute_actor`
has been updated to use this new flag.

This preserves the existing behavior and memory usage while allowing for
up to 7 additional flags to be added to `actor->flags` if needed.