Upgrade Alpine version used for CI to 3.16
We were using 3.12 which reached end of life on May 1st of this year.
This gets up to using the newest alpine release that will be supported
until May of 2024.

If this all passes CI and stress tests run as usual, I'll open a
PR to update the user facing Alpine based images.