Actor should be unmuted even if it is unscheduled
Currently, an actor is not unmuted if it is unscheduled. This is an
issue because `pony_schedule` will not reschedule an actor if it is
currently muted. The current logic would result in unscheduled and
muted actors never being unmuted and rescheduled.

This commit changes the logic in `ponyint_sched_unmute_senders` to
always unmute actors that should be unmuted but to only redschedule
them if they were not unscheduled.