Remove `FLAG_UNSCHEDULED` and related code (#4137)
Per @SeanTAllen, `library` mode has been removed but some supporting
code is pending removal. The `FLAG_UNSCHEDULED` and related logic is
some of that code that is pending removal.

This commit removes `FLAG_UNSCHEDULED` and related code.

`pony_become` was needed for `library` mode which
is no longer supported and was referenced by `FLAG_UNSCHEDULED` code.

This commit renames `pony_become` to `ponyint_become` and removes it
from the `PONY_API`. References in codegen and the cycle detector have
been updated as necessary.