Remove `ponyint_heap_alloc*_final` functions
Prior to this commit, there were a number of heap allocation
functions for finalisers where the only difference from the
normal heap allocation functions was that the finaliser
functions kept track of which finalisers needed to be run
in the `chunk->finalisers` bitmap.

This commit combines the finaliser heap allocation functions
with their counterpart normal heap allocations functions.
It accomplishes this by adding in an extra argument called
`track_finalisers_mask` that can only be `0xFFFFFFFF` to
signify that finalisers should be tracked or `0` to signify
that finalisers shouldn't be tracked via a branchless bitwise
AND operation. This lowers the maintenance burden as now a
change to heap allocation logic will not need to be made
in two places at the same time and might even result in a
tiny performance increase due to the ever so slightly smaller
binary size.