Avoid clearing chunks at start of GC
Prior to this commit, at the start of GC, the heap clears all
`chunks` by setting them all to `empty` in `ponyint_heap_startgc`.
This is an expensive operation as it requires visiting all
`chunks` and updating their `chunk->slots` and `chunk->shallow`
values regardless of whether they are still in use or not.
`chunk->shallow` is only ever used as part of the GC process.

This commit uses the `chunk->shallow` with a `sentinel`
value to avoid clearing chunks for every sizeclass except
for sizeclass 0 which is still cleared and makes sure that
the mark and sweep steps clear the chunk if
`chunk->shallow` has the `sentinel` value and needs to be
cleared. Additionally, `ponyint_heap_ismarked` is removed
as it is not used.