Enhance runtime memory allocation tracking
Runtime allocation tracking now also tracks the number
of heap allocations, the number of freed heap
allocations and the number of GC iterations via counters.

Additionally, there is now a way to check if runtime
memory allocation tracking is enabled or not via
`ifdef` statements in Pony code. This allows for some
useful validations for those folks concerned about
heap allocations in the critical path (i.e. if they
rely on the compiler's `HeapToStack` optimization pass
to convert heap allocations to stack allocations and
want to validate it is working correctly).