Ensure reachability of types returned by FFI calls
After introducing #4060, we forgot to re-add the code that ensured that, if a
FFI function call specified different return types, those types were being
marked as reachable.

This created a situation where programs that introduced a new type (for example,
a bare lambda) in the context of a FFI return type would crash the compiler.
This was due to having missing information about the type during code
generation.