Auto-recover constructor expressions (#4124)
This PR adds checks when assigning a constructor expression or passing one as a parameter to see if it can be auto-recovered, and if so, does not output a subtyping error. E.g. for the following code:

```
actor Main
  new create(env: Env) =>
    Bar.take_foo(Foo(88))
    let bar: Foo iso = Foo(88)

class Foo
  new create(v: U8) =>
    None

primitive Bar
  fun take_foo(foo: Foo iso) =>
    None
```

We'd previously get errors on both lines in Main.create(); whereas with this PR they are now accepted.

Fixes #702


Co-authored-by: Sean T Allen <sean@seantallen.com>
Co-authored-by: Jason Carr <jcarr250@protonmail.com>