Update windows ci image (#4170)
This change updates the Windows build image to use the latest Visual Studio Build Tools, as well as the latest Python 3.

In addition, it adds an option for building LLDB on Windows as part of `make libs`.  This is to support #4164.  Prebuilt binaries via Chocolatey or otherwise don't seem to work right on X64 Windows, so we need to build it ourselves.  Per https://github.com/llvm/llvm-project/issues/54555#issuecomment-1079312120 we also have to build Clang as well.