Run test suites and stress test in a debugger (#4164)
We currently sometimes get segfaults and other crashes when running
our test suites and stress tests via CI. But we don't get any useful
information when these crashes occur.

This commit adds gdb to most of our docker containers and uses either
gdb or lldb to run the test suites and stress tests in CI with
appropriate arguments to print out backtraces, locals, arguments, and
registers prior to quitting with error. While the backtraces aren't
always useful (due to optimizations or lack of symbols), they are
better than nothing.

lldb is used for all containers except for x86 musl which uses gdb.
We use gdb for that container as for some reason, using lldb is
resulting in test hangs that we haven't explored.