Only run libponyc.run.tests in debug mode
Currently for both release ande debug runtimes, we compile the test
programs both with and without --debug. This commit changes to only
build with --debug.

This change is because currently tests are taking a really long time
with the release version of the runtime. This is a hopefully temporary
fix to speed things up.