Fix broken readline package (#4199)
With the release of 0.49.0, we changed `Ansi.erase` to be erasing left
rather than right. This made sense based on the comment on the `Ansi.erase`
method. However, the actual usage in the standard library of `Ansi.erase`
was expecting it to erase to the right. Because of this change, since
0.49.0, the readline support in the `term` package has been broken.

This commit switches erase back to erasing to the right and in the process
fixes the readline package. The commment on `erase` has been updated to
reflect what it actually does.

Supporting erasing to both the left and the right seems valuable but, I'm
not addressing that here and instead am leaving that to an RFC.