Split Windows runner tests apart in time
Through testing, it seems like running the runner tests back to back
ends up with us hitting problems where the machine seems to hang.

From some initial testing it looks like splitting them apart so
they run with a gap between them might solve the issue. More testing
is needed but if this change passes another bit of testing it
is probably worth trying this out for a bit and see if it continues
to make the tests run and not hang.