Set Windows runner tests to be run sequentially
The runner relies on a timer to start each of the runner tests when
not run sequentially. We are seeing the Windows runs hang. This
could be explained by us occassionally losing IOCP events.

If the problem goes away when running sequentially and stays away
for a while we can rewrite the parallel test run code to sidestep
the problem and verify the issue. If we are losing events, it
would explain issue #3981 and the flakey Windows network tests.