Strength the order for an atomic operation in the IOCP subsystem (#4213)
In PR #4083, I reversed some relaxing of atomic orderings that were
introduced in 2016 in #1209. When I was doing that change, I missed
an additional strengthing that should have been done in the IOCP
subsystem.

If you note the original commit, I made a corresponding change in
the EPoll subsystem, but missed doing it in the IOCP system.