Change libponyc.run test runner parallelism strategy
We've been encountering issues with Windows libponyc.run.tests hanging. Or rather, with what appears to be
the test runner itself hanging. 

After a lot of playing around, I determined that running the tests using the `--sequential` option appears to fix the problem as I was able run several times without issue. Given previous experience, during that time, I would have
expected a hang if the problem still existed.

The key difference between the testing with and without `--sequential` was that without `--sequential`, a timer is used to start each of the runner tests. If we were sometimes losing timer events then eventually, we might stop running test entirely and the runner would hang without exiting.

This commit changes the implementation of parallel test running to not use a timer. Really, there was no need for a timer. We can start our X number of tests and then as each completes, start a corresponding new test until we run
out of tests to run.

Testing indicates that this change fixes our Windows hanging issue. We might not in fact have been losing timer events, there's no proof of that. I only have some level of proof that with this change, the problem appears to be fixed. Any thoughts about the IOCP subsystem or something else "losing" timer events is mostly conjecture based on what I have seen. At this time, I'm willing to accept this set of changes as I think it is a better overall approach than the timer based on and it appears to fix our hanging problem.

As part of this change, `--sequential` has been removed as an option. To run tests in a sequential fashion, use a parallelism value of `1`. Additionally, testing showed that running tests in parallel on non-Windows platforms leads to a decrease in the time taken to run the tests. There is no corresponding decrease in time to finish the tests on Windows and so, after testing that the change in how parallel running is implemented on Windows, I set the parallelism to one which mirrors the old `--sequential` option.