Adapt docgen to use the mkdocs material theme
As the ponylang-mkdocs-theme wasn't maintainable anymore.
This puts all the necessary tweaks inside the generated files (logo, custom css, theme adaptions and extensions)
and only depends on the user to have mkdocs and the python package material-mkdocs installed to generate and serve html docs.

This also contains an improvement to sort items in a package alphabetically to make it easier to find classes/actors/primitives.