Update docgen to generate source files that are ignored (#4239)
This feature is currently "sponsors only" so it won't work for
regular users. However, we can switch our Pony projects over to
using the insiders theme and take advantage of this change.

This will improve the search experience by not including our src
files in the search index.

Any users that use mkdocs-material-insiders will also be able to take
advantage of this change.