Use mkdocs-material-insiders theme when manually building stdlib documentation
This is a stepping stone to having private images that are for ponylang org
only usage for building our various bits of documentation that contain the
theme in addition to images for public consumption that don't contain the
insiders theme.

If this test works out, I'll start working on rolling out.