Install mkdocs-material-insiders in standard library builder image (#4241)
In theory, this is a simple thing to do. However, there's at least a
couple additional moving parts here because, we do not want to expost
squidfunk's "private insiders code" in a public image on docker hub.

The commit changes our standard library builder image to live in the
GitHub container registry where we have marked it as private so only
we can access it. By having the builder as private, we can safely install
the insiders theme in it and use it for building our standard library
docs.

To install in the image, we need to supply the correct access token as
part of the build process. The token has been stored in a Ponylang org
secret which is supplied at the time we build on GitHub. Because this
makes building locally from the command line more complicated, we have
also added a workflow_dispatch workflow to rebuild the stdlib-builder in
an adhoc fashion via the GitHub UI if an adhoc rebuild is needed.