Don't include "home file" in documentation search
The "home file" is only a list of the packages on the site. It ends
up being extra noise in search results where it provides no value.

This commit adds front matter to the home file that will remove
it from the search index when the mkdocs-material-insiders theme
is removed.