Make set/unset of actor sync flags atomic
Currently, each is a read and then a write and this is in theory
completely fine as we will only ever write from a single thread
to the flags (the thread running the actor) and then might read
from other threads.

I don't see any reason for the above theory to be disproven, however,
I find myself somewhat more at ease with doing a simple atomic operation
rather than two for set and unset.