Make the full-program test runner able to run tests in a debugger
Adds a `--debugger` option to the runner command line, which takes a debugger command line with the executable and args.  If this is a non-empty string, the runner will run tests using that debugger and args.

Also parses LLDB output to get the test's exit code, since LLDB will always exit with 0.

Also adds the `--debugger` option to the Makefile and make.ps1 to use the debugger if configured.

Fixes #4222